Fix all failing tests
