+++ .ruby-version
@@ -1,1 +1,1 @@
-ruby-2.7.0
+ruby-3.2.2
