+++ app/views/posts/index.html.erb
@@ -5,7 +5,7 @@
 <div>
   <div class="parent-container">
     <div class="user-image">
-      <img src=<%= @user.photo %> width="75rem">
+      <img src=<%= @user.photo %> width="75rem" />
     </div>
 
     <div class="user-details">
