+++ app/views/users/index.html.erb
@@ -6,7 +6,7 @@
   <%@users.each do |user|%>
   <div class="parent-container">
     <div class="user-image">
-      <img src=<%= user.photo %> width="75rem">
+      <img src=<%= user.photo %> />
     </div>
 
     <div class="user-details">
