+++ config/routes.rb
@@ -4,6 +4,7 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   # root "articles#index"
   root 'users#index'
+  get '/www.eample.com', to: 'users#index'
   
   resources :users, only: [:index, :show] do
     resources :posts, only: [:index, :show, :new, :create] do
