+++ spec/integration/user_show_spec.rb
@@ -11,29 +11,6 @@ RSpec.describe 'Users', type: :feature do
                             likes_counter: 0)
   end
 
-  describe 'Test for index page of users' do
-    it 'should test for correct content on the page' do
-      visit users_path
-      expect(page).to have_content(@user.name)
-    end
-
-    it 'should show profile picture for each user' do
-      visit users_path
-      expect(page).to have_css('img[src="https://unsplash.com/photos/F_-0BxGuVvo"]')
-    end
-
-    it 'should show number of posts for each user' do
-      visit users_path
-      expect(page).to have_content(@user.posts_counter)
-    end
-
-    it 'should redirect when i click on user to shoe page' do
-      visit users_path
-      click_link @user.name
-      expect(page).to have_current_path(user_path(@user))
-    end
-  end
-
   describe 'Test for show page for users' do
     it 'should show users profile picture' do
       visit user_path(@user.id)
